requirements.txt
pillow
beautifulsoup4  
